Rename: .swift 파일, 폴더 재정립
- ViewModel, Singleton 폴더 변경 후 CacheManager, WebService, BooksSearchViewModel 파일 이동
- View extension은 View+로 네이밍 변경
- CustomUI는 폴더 생성 후 별도 분리
- SearchBooksResult 파일 생성 후 별도 모델 관리
