+++ BookSearchApp/BookSearchApp/Architecture/Singleton/WebService.swift
